requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.6
paho-mqtt==1.5.0
pika==1.3.2
pymongo==4.6.2
requests==2.31.0
urllib3==2.2.1
